+++ ESP32/ESP_Pesagem_Final/lib/can_info/src/HX711.c
@@ -12,7 +12,7 @@ static gpio_num_t GPIO_PD_SCK = GPIO_NUM_15;	// Power Down and Serial Clock Inpu
 static gpio_num_t GPIO_DOUT = GPIO_NUM_18;		// Serial Data Output Pin
 static HX711_GAIN GAIN = eGAIN_128;		// amplification factor
 static unsigned long OFFSET = 0;	// used for tare weight
-static float SCALE = 1;	// used to return weight in grams, kg, ounces, whatever
+static double SCALE = 1;	// used to return weight in grams, kg, ounces, whatever
 
 void HX711_init(gpio_num_t dout, gpio_num_t pd_sck, HX711_GAIN gain )
 {
@@ -129,7 +129,7 @@ unsigned long HX711_get_value(char times)
 		return 0;
 }
 
-float HX711_get_units(char times) 
+double HX711_get_units(char times) 
 {
 	return HX711_get_value(times) / SCALE;
 }
@@ -143,12 +143,12 @@ void HX711_tare( )
 	ESP_LOGI(DEBUGTAG, "===================== END TARE: %ld ====================",sum);
 }
 
-void HX711_set_scale(float scale ) 
+void HX711_set_scale(double scale ) 
 {
 	SCALE = scale;
 }
 
-float HX711_get_scale()
+double HX711_get_scale()
  {
 	return SCALE;
 }
